refactor(stage-2a): extract AI advisor controller; eliminate module-scope run state
Per docs/plans/stages/stage-2a-advisor-controller.md:
- New ai/advisorController.ts: askForMove / toggleAutoPlay / continueAutoPlay
  / cancel moved verbatim from the store; receives {get, set, engine} by
  injection so ai/ still never imports the store
- The 7 module-scope mutable variables (abort controller, loop/stall
  tracking, turn cap) become controller-owned closure state with an explicit
  resetRunState() — initializeGame and loadSavedSession now call it instead
  of hand-clearing a subset
- Owned timer handle: the scheduled next-turn / retry-cooldown setTimeout is
  cleared on stop and reset (previously orphaned, benign only by re-check)
- exportAIInteractions reads the stall outcome via advisor.wasStalled()
- New controller-level tests (timer cleared on reset/stop; no-key start
  opens the key modal and disarms)
- gameStore.ts: 1,746 -> 1,205 lines; zero module-scope 'let' remaining

App suite 378 green (aiAdvisor.test.ts unmodified); ai-advisor e2e (11) green.
